Moidfy

+++ dto/systemConfig_dto.go
@@ -6,7 +6,7 @@ type CreateSystemConfigDto struct {
 }
 
 type UpdateSystemConfigDto struct {
-	ID                    uint   `json:"id" form:"id" binding:"required"`
+	ID                    uint64 `json:"id" form:"id" binding:"required"`
 	TeacherCanBorrowCount uint64 `json:"teacher_can_borrow_count" form:"teacher_can_borrow_count" binding:"required"`
 	StudentCanBorrowCount uint64 `json:"student_can_borrow_count" form:"student_can_borrow_count" binding:"required"`
 }
