+++ model/punishment.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Punishment struct {
-	ID           uint           `gorm:"column:id;primary_key:auto_increment" json:"id"`
+	ID           uint64         `gorm:"column:id;primary_key:auto_increment" json:"id"`
 	PackageName  string         `gorm:"column:package_name;type:varchar(125)" json:"package_name"`
 	Duration     uint64         `gorm:"column:duration;" json:"duration"`
 	PunishAmount uint64         `gorm:"column:punish_amount;not null" json:"punish_amount"`
